Implemented make variable, renamed structs
MakeVariable is a function that turns a given value into a variable and assigns it a proper type. This is used instead of the constructors when you want a responsible variable, and not a value.

renamed GetVariableType and GetReturnType to GetValueType and GetStorageType to reflect what is being gotten more effectively.