introduced ref accounting system
Introduced a reference accounting system to tell if a runtime variable is being referenced anywhere, as well as telling when a referenced variable is being deleted.

Introduced clear to runtime variable, instead of using clear in variable. Solves a bug where a ref rv would be null when gotten.

Introduced AssignRef, that uses similar checks to Variable::Assign, but is designed for trading rvs.

Note: AssignRef currently passes all queries.